+++ src/main/java/nextstep/courses/infrastructure/JdbcSessionRepository.java
@@ -1,6 +1,6 @@
 package nextstep.courses.infrastructure;
 
-import nextstep.courses.domain.cource.ImageRepository;
+import nextstep.courses.domain.session.ImageRepository;
 import nextstep.courses.domain.session.Session;
 import nextstep.courses.domain.session.SessionRepository;
 import nextstep.courses.domain.session.StudentsRepository;
