+++ src/test/java/nextstep/courses/domain/session/SessionTest.java
@@ -1,6 +1,5 @@
 package nextstep.courses.domain.session;
 
-import nextstep.courses.domain.cource.Image;
 import nextstep.courses.exception.session.EnrollmentMaxExceededException;
 import nextstep.courses.exception.session.InvalidPaymentAmountException;
 import nextstep.courses.exception.session.InvalidSessionStateException;
